+++ pkg/networkserver/grpc_gsns.go
@@ -50,12 +50,6 @@ const (
 	// recentUplinkCount is the maximum amount of recent uplinks stored per device.
 	recentUplinkCount = 20
 
-	// retransmissionWindow is the maximum delay between Rx2 end and an uplink retransmission.
-	retransmissionWindow = 10 * time.Second
-
-	// maxConfNbTrans is the maximum number of confirmed uplink retransmissions for pre-1.0.4 devices.
-	maxConfNbTrans = 5
-
 	// joinRequestCollectionWindow is the duration for which duplicated JoinRequests are collected.
 	// This parameter is separated from the uplink collection period since the JoinRequest may have to be
 	// served by a Join Server which is either geographically far away, or simply slow to respond.
@@ -91,18 +85,14 @@ func (ns *NetworkServer) deduplicateUplink(
 	return true, nil
 }
 
-func maxTransmissionNumber(ver ttnpb.MACVersion, confirmed bool, nbTrans uint32) uint32 {
+func (ns *NetworkServer) maxTransmissionNumber(ver ttnpb.MACVersion, confirmed bool, nbTrans uint32) uint32 {
 	if !confirmed {
 		return nbTrans
 	}
 	if macspec.LimitConfirmedTransmissions(ver) {
-		return maxConfNbTrans
+		return ns.maxConfNbTrans
 	}
 	return nbTrans
-}
-
-func maxRetransmissionDelay(rxDelay ttnpb.RxDelay) time.Duration {
-	return rxDelay.Duration() + time.Second + retransmissionWindow
 }
 
 func matchCmacF(ctx context.Context, fNwkSIntKey types.AES128Key, macVersion ttnpb.MACVersion, fCnt uint32, up *ttnpb.UplinkMessage) ([4]byte, bool) {
@@ -376,7 +366,11 @@ func (ns *NetworkServer) matchAndHandleDataUplink(ctx context.Context, dev *ttnp
 			default: // cmacFMatchResult.FullFCnt == dev.Session.LastFCntUp
 				ctx = log.NewContextWithField(ctx, "f_cnt_reset", false)
 
-				maxNbTrans := maxTransmissionNumber(dev.MacState.LorawanVersion, up.Payload.MHdr.MType == ttnpb.MType_CONFIRMED_UP, dev.MacState.CurrentParameters.AdrNbTrans)
+				maxNbTrans := ns.maxTransmissionNumber(
+					dev.MacState.LorawanVersion,
+					up.Payload.MHdr.MType == ttnpb.MType_CONFIRMED_UP,
+					dev.MacState.CurrentParameters.AdrNbTrans,
+				)
 				if maxNbTrans < 1 {
 					panic(fmt.Sprintf("invalid maximum transmission number %d", maxNbTrans))
 				}
@@ -414,19 +408,10 @@ func (ns *NetworkServer) matchAndHandleDataUplink(ctx context.Context, dev *ttnp
 					log.FromContext(ctx).Debug("Repeated FCnt value, but frame is not a retransmission")
 					return nil, false, nil
 				}
-				maxDelay := maxRetransmissionDelay(dev.MacState.CurrentParameters.Rx1Delay)
-				delay := ttnpb.StdTime(up.ReceivedAt).Sub(lastAt)
 				ctx = log.NewContextWithFields(ctx, log.Fields(
 					"last_transmission_at", lastAt,
-					"max_retransmission_delay", maxDelay,
-					"retransmission_delay", delay,
 					"retransmission_number", nbTrans,
 				))
-				if delay > maxDelay {
-					log.FromContext(ctx).Warn("Retransmission delay exceeds maximum")
-					return nil, false, nil
-				}
-
 				trace.Log(ctx, "ns", "current session match with retransmission")
 				matchType = currentRetransmissionMatch
 			}