+++ CHANGELOG.md
@@ -13,9 +13,13 @@ For details about compatibility between different releases, see the **Commitment
 
 ### Changed
 
+- Increased the default value from 5 to 15 for maximum number of confirmed uplink retransmission for LoRaWAN version prior 1.0.4.
+
 ### Deprecated
 
 ### Removed
+
+- Maximum retransmission delay window for confirmed uplinks.
 
 ### Fixed